+++ src/Table/index.stories.tsx
@@ -13,8 +13,12 @@ export default {
     size: {
       control: {
         type: 'select',
-        options: ['small', 'medium', 'large'],
+        options: ['small', 'middle', 'large'],
       },
+    },
+    loading: {
+      defaultValue: false,
+      control: 'boolean',
     },
   },
 };