+++ src/theme.ts
@@ -1,3 +1,4 @@
+import { SizeType } from 'antd/es/config-provider/SizeContext';
 import { useContext } from 'react';
 import { ThemeContext } from 'styled-components';
 
@@ -34,7 +35,31 @@ const DEPRECATED_PALETTE = {
 };
 /* eslint-enable @mll-lab/no-color-literals */
 
-export const THEME = {
+export type Theme = {
+  backgroundColor: string;
+  borderColor: string;
+  collapseBackgroundColor: string;
+  containerBorderColor: string;
+  dividerColor: string;
+  focusedRowColor: string;
+  menuGroupBackgroundColor: string;
+  panelBackgroundColor: string;
+  tableBorderColor: string;
+  titleColor: string;
+  disabledColors?: {
+    lowContrast: string;
+  };
+
+  successColor: string;
+  warningColor: string;
+  errorColor: string;
+  infoColor: string;
+
+  fontSize?: string;
+  size?: SizeType;
+};
+
+export const THEME: Theme = {
   // Components
   backgroundColor: PALETTE.blueTintedGray,
   borderColor: PALETTE.blue,
@@ -46,15 +71,19 @@ export const THEME = {
   panelBackgroundColor: DEPRECATED_PALETTE.gray1,
   tableBorderColor: PALETTE.gray3,
   titleColor: PALETTE.gray5,
+  disabledColors: {
+    lowContrast: PALETTE.gray5,
+  },
 
   // UI states
   successColor: PALETTE.green,
   warningColor: PALETTE.gold,
   errorColor: PALETTE.red,
-  infoColor: PALETTE.gray4,
-};
+  infoColor: PALETTE.gray7,
 
-export type Theme = typeof THEME;
+  // Sizes
+  fontSize: '14px', // antd default
+};
 
 export function useTheme(): Theme {
   return useContext(ThemeContext) ?? THEME;