+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [7.8.0](https://github.com/mll-lab/react-components/compare/v7.7.0...v7.8.0) (2022-01-04)
+
+
+### Features
+
+* mll logo as default table loading indicator ([#116](https://github.com/mll-lab/react-components/issues/116)) ([c5d4cba](https://github.com/mll-lab/react-components/commit/c5d4cba101c78f62026d085226c192fd5d7b0fae))
+
+# [7.7.0](https://github.com/mll-lab/react-components/compare/v7.6.0...v7.7.0) (2022-01-04)
+
+
+### Features
+
+* distinct disabled button style ([#111](https://github.com/mll-lab/react-components/issues/111)) ([0e7e495](https://github.com/mll-lab/react-components/commit/0e7e495792e61ababa1a708972b25628fd3cb944))
+
+# [7.6.0](https://github.com/mll-lab/react-components/compare/v7.5.0...v7.6.0) (2022-01-03)
+
+
+### Features
+
+* provide component size and font-size ([#113](https://github.com/mll-lab/react-components/issues/113)) ([64705c6](https://github.com/mll-lab/react-components/commit/64705c69804c4e06ae010d8d4433d3d292c23f8e))
+
 # [7.5.0](https://github.com/mll-lab/react-components/compare/v7.4.1...v7.5.0) (2021-12-20)